// packages/plugin-calendar-ui/src/settings/types.ts
export interface ICalendar {
  _id: string;
  name: string;
  color: string;
  groupId: string;
  accountId: string;
  userId: string;
}
// ...
export type CalendarsQueryResponse = {
  calendars: ICalendar[];
  loading: boolean;
};
// ...
export type RemoveSchedulePageMutationResponse = {
  removeMutation: (params: {
    variables: { pageId: string };
    refetchQueries?: any;
  }) => Promise<void>;
};